Revert "SSCSSI-247: Merged with master"
This reverts commit 90a1390b7b060c24c3e51a5f5f68ee56d1b7ff1f.
