+++ src/main/java/uk/gov/hmcts/reform/sscs/service/SubmitAppealService.java
@@ -11,7 +11,6 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.tuple.Pair;
 import org.apache.http.HttpStatus;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -159,24 +158,27 @@ public class SubmitAppealService extends SubmitAppealServiceBase {
 
     private SscsCaseDetails createOrUpdateCase(SscsCaseData caseData, EventType eventType, IdamTokens idamTokens) {
         SscsCaseDetails caseDetails = null;
-        String benefitShortName = caseData.getAppeal().getBenefitType().getCode();
-        String nino = caseData.getAppeal().getAppellant().getIdentity().getNino();
-        String ibcaReference = caseData.getAppeal().getAppellant().getIdentity().getIbcaReference();
+
         try {
-            Pair<List<SscsCaseDetails>, SscsCaseDetails> matchedByNinoCasesCaseDetails =
-                    getMatchedNinoCasesAndCaseDetails(caseData, idamTokens, nino, ibcaReference);
-            List<SscsCaseDetails> matchedByNinoCases = matchedByNinoCasesCaseDetails.getLeft();
-            caseDetails = matchedByNinoCasesCaseDetails.getRight();
+            List<SscsCaseDetails> matchedByNinoCases = getMatchedCases(caseData.getAppeal().getAppellant().getIdentity().getNino(), idamTokens);
+            if (!matchedByNinoCases.isEmpty()) {
+                log.info("Found " + matchedByNinoCases.size() + " matching cases for Nino "
+                        + caseData.getAppeal().getAppellant().getIdentity().getNino() + " before filtering non exact matches");
+            } else {
+                log.info("No matching cases for Nino {}", caseData.getAppeal().getAppellant().getIdentity().getNino());
+            }
+            caseDetails = matchedByNinoCases.stream().filter(createNinoAndBenefitTypeAndMrnDatePredicate(caseData)).findFirst().orElse(null);
+
             if (caseDetails == null) {
                 if (!matchedByNinoCases.isEmpty()) {
                     log.info("Found " + matchedByNinoCases.size() + " matching cases for Nino "
-                            + maskNino(nino));
+                            + caseData.getAppeal().getAppellant().getIdentity().getNino());
                     addAssociatedCases(caseData, matchedByNinoCases);
                 }
 
                 log.info("About to attempt creating case or updating draft case in CCD with event {} for benefit type {} and event {} and isScottish {} and languagePreference {}",
                         eventType,
-                        benefitShortName,
+                        caseData.getAppeal().getBenefitType().getCode(),
                         eventType,
                         caseData.getIsScottishCase(),
                         caseData.getLanguagePreference().getCode());
@@ -193,7 +195,7 @@ public class SubmitAppealService extends SubmitAppealServiceBase {
 
                     log.info("Case {} successfully converted from Draft to SSCS case in CCD for benefit type {} with event {}",
                             caseDetails.getId(),
-                            benefitShortName,
+                            caseData.getAppeal().getBenefitType().getCode(),
                             eventType);
                 } else {
                     caseDetails = ccdService.createCase(caseData,
@@ -203,26 +205,23 @@ public class SubmitAppealService extends SubmitAppealServiceBase {
                             idamTokens);
                     log.info("Case {} successfully created in CCD for benefit type {} with event {}",
                             caseDetails.getId(),
-                            benefitShortName,
+                            caseData.getAppeal().getBenefitType().getCode(),
                             eventType);
                 }
                 return caseDetails;
             }
         } catch (Exception e) {
-            String caseId = caseDetails != null ? caseDetails.getId().toString() : "";
-            String referenceName = caseData.isIbcCase() ? "IBCA Reference" : "Nino";
-            String referenceValue = caseData.isIbcCase() ? ibcaReference : maskNino(nino);
             throw new CcdException(
                     String.format("Error found in the creating case process for case with Id - %s"
-                                    + " and %s - %s and Benefit type - %s and exception: %s",
-                            caseId, referenceName, referenceValue,
-                            benefitShortName, e.getMessage()), e);
+                                    + " and Nino - %s and Benefit type - %s and exception: %s",
+                            caseDetails != null ? caseDetails.getId() : "", caseData.getAppeal().getAppellant().getIdentity().getNino(),
+                            caseData.getAppeal().getBenefitType().getCode(), e.getMessage()), e);
         }
 
         log.info("Duplicate case {} found for Nino {} and benefit type {}. "
                         + "No need to continue with post create case processing.",
-                caseDetails.getId(), maskNino(nino),
-                benefitShortName);
+                caseDetails.getId(), caseData.getAppeal().getAppellant().getIdentity().getNino(),
+                caseData.getAppeal().getBenefitType().getCode());
         throw new DuplicateCaseException(
                 String.format("An appeal has already been submitted, for that decision date %s ",
                         caseData.getAppeal().getMrnDetails().getMrnDate()));
