+++ src/main/java/uk/gov/hmcts/reform/sscs/service/v2/SubmitAppealService.java
@@ -14,7 +14,6 @@ import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.function.UnaryOperator;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.tuple.Pair;
 import org.apache.http.HttpStatus;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -183,24 +182,25 @@ public class SubmitAppealService extends SubmitAppealServiceBase {
     }
 
     private SscsCaseDetails createOrUpdateCase(SyaCaseWrapper appeal, SscsCaseData caseData, EventType eventType, IdamTokens idamTokens) {
+        Optional<SscsCaseDetails> optionalSscsCaseDetails;
         SscsCaseDetails caseDetails = null;
-        String benefitShortName = caseData.getAppeal().getBenefitType().getCode();
-        String nino = caseData.getAppeal().getAppellant().getIdentity().getNino();
-        String ibcaReference = caseData.getAppeal().getAppellant().getIdentity().getIbcaReference();
         try {
-            Pair<List<SscsCaseDetails>, SscsCaseDetails> matchedByNinoCasesCaseDetails =
-                    getMatchedNinoCasesAndCaseDetails(caseData, idamTokens, nino, ibcaReference);
-            List<SscsCaseDetails> matchedByNinoCases = matchedByNinoCasesCaseDetails.getLeft();
-            caseDetails = matchedByNinoCasesCaseDetails.getRight();
-            if (caseDetails == null) {
-                if (!matchedByNinoCases.isEmpty()) {
-                    log.info("Found " + matchedByNinoCases.size() + " matching cases for Nino "
-                            + maskNino(nino));
-                }
+            List<SscsCaseDetails> matchedByNinoCases = getMatchedCases(caseData.getAppeal().getAppellant().getIdentity().getNino(), idamTokens);
+            if (!matchedByNinoCases.isEmpty()) {
+                log.info("Found {} matching cases for Nino {} before filtering non exact matches", matchedByNinoCases.size(),
+                        caseData.getAppeal().getAppellant().getIdentity().getNino());
+            } else {
+                log.info("No matching cases for Nino {}", caseData.getAppeal().getAppellant().getIdentity().getNino());
+            }
 
+            optionalSscsCaseDetails = matchedByNinoCases.stream()
+                    .filter(createNinoAndBenefitTypeAndMrnDatePredicate(caseData))
+                    .findFirst();
+
+            if (optionalSscsCaseDetails.isEmpty()) {
                 log.info("About to attempt creating case or updating draft case in CCD with event {} for benefit type {} and event {} and isScottish {} and languagePreference {}",
                         eventType,
-                        benefitShortName,
+                        caseData.getAppeal().getBenefitType().getCode(),
                         eventType,
                         caseData.getIsScottishCase(),
                         caseData.getLanguagePreference().getCode());
@@ -240,20 +240,17 @@ public class SubmitAppealService extends SubmitAppealServiceBase {
                 return caseDetails;
             }
         } catch (Exception e) {
-            String caseId = caseDetails != null ? caseDetails.getId().toString() : "";
-            String referenceName = caseData.isIbcCase() ? "IBCA Reference" : "Nino";
-            String referenceValue = caseData.isIbcCase() ? ibcaReference : maskNino(nino);
             throw new CcdException(
                     String.format("Error found in the creating case process for case with Id - %s"
-                                    + " and %s - %s and Benefit type - %s and exception: %s",
-                            caseId, referenceName, referenceValue,
-                            benefitShortName, e.getMessage()), e);
+                                    + " and Nino - %s and Benefit type - %s and exception: %s",
+                            caseDetails != null ? caseDetails.getId() : "", caseData.getAppeal().getAppellant().getIdentity().getNino(),
+                            caseData.getAppeal().getBenefitType().getCode(), e.getMessage()), e);
         }
-
+        caseDetails = optionalSscsCaseDetails.get();
         log.info("Duplicate case {} found for Nino {} and benefit type {}. "
                         + "No need to continue with post create case processing.",
-                caseDetails.getId(), maskNino(nino),
-                benefitShortName);
+                caseDetails.getId(), caseData.getAppeal().getAppellant().getIdentity().getNino(),
+                caseData.getAppeal().getBenefitType().getCode());
         throw new DuplicateCaseException(
                 String.format("An appeal has already been submitted, for that decision date %s ",
                         caseData.getAppeal().getMrnDetails().getMrnDate()));
